requirements.txt
fastapi
opencv-python
mrz-scanner-sdk
numpy
